+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] / 2022-05-07
+### Features
+- IsLanguageExtension for each Language
+- Remove RibbonTab name from Revit Dictionary - `GetRibbonTabsDictionary`
+### Added
+- Add `Autodesk.Windows.ComponentManager.IsApplicationFrameEnabled` to disable buttons on modeless Example
+
 ## [0.1.0] / 2022-04-04
 ### Added
 - Add `revit.ico` to Test
@@ -140,6 +147,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.1.1]: ../../compare/0.1.0...0.1.1
 [0.1.0]: ../../compare/0.0.9...0.1.0
 [0.0.9]: ../../compare/0.0.8...0.0.9
 [0.0.8]: ../../compare/0.0.7...0.0.8